@@ -1,19 +1,19 @@
 source 'http://rubygems.org'
 
-# Rails
-gem 'rails', '3.1.0'
-gem 'rake', '0.9.2'
+gem 'rails', '3.2.2'
 
-# MySQL Driver
-gem 'pg'
+# Bundle edge Rails instead:
+# gem 'rails', :git => 'git://github.com/rails/rails.git'
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails', "  ~> 3.1.0"
-  gem 'coffee-rails', "~> 3.1.0"
-  gem 'uglifier'
+  gem 'sass-rails', "~> 3.2.3"
+  gem 'coffee-rails', "~> 3.2.1"
+  gem 'uglifier', '>= 1.0.3'
 end
+
+gem 'pg'
 
 # Javascript stuff.
 gem 'json'
@@ -21,6 +21,7 @@ gem 'jquery-rails'
 
 # For file attachments.
 gem 'paperclip'
+gem 'aws-s3'
 
 # Bitly
 gem 'bitly'
