@@ -13,17 +13,17 @@ class ApplicationController < ActionController::Base
       errors << error[0].to_s.titleize + " " + error[1]
     end
 
-    return errors
+    errors
   end
 
 
   protected
   def authenticate
     if params[:auth] && params[:auth][:username] == "com.medlmobile.vosn" && params[:auth][:password] == "6381d1297a6e9b54a19c529ea366c32b"
-      return true
+      true
     else
       head :unauthorized
-      return false
+      false
     end
   end
 end
